+++ src/main/java/onlyoffice/sdk/service/config/ConfigServiceImpl.java
@@ -26,7 +26,11 @@ import com.onlyoffice.manager.security.JwtManager;
 import com.onlyoffice.manager.settings.SettingsManager;
 import com.onlyoffice.manager.url.UrlManager;
 import com.onlyoffice.model.common.User;
+import com.onlyoffice.model.documenteditor.Config;
 import com.onlyoffice.model.documenteditor.config.document.Permissions;
+import com.onlyoffice.model.documenteditor.config.document.Type;
+import com.onlyoffice.model.documenteditor.config.editorconfig.Mode;
+import com.onlyoffice.model.documenteditor.config.editorconfig.customization.Anonymous;
 import com.onlyoffice.service.documenteditor.config.DefaultConfigService;
 import onlyoffice.utils.AttachmentUtil;
 
@@ -42,6 +46,23 @@ public class ConfigServiceImpl extends DefaultConfigService {
     }
 
     @Override
+    public Config createConfig(final String fileId, final Mode mode, final Type type) {
+        Config config = super.createConfig(fileId, mode, type);
+
+        if (mode.equals(Mode.VIEW) || config.getDocument().getPermissions().getEdit().equals(false)) {
+            Anonymous anonymous = Anonymous.builder()
+                    .request(false)
+                    .build();
+
+            config.getEditorConfig()
+                    .getCustomization()
+                    .setAnonymous(anonymous);
+        }
+
+        return config;
+    }
+
+    @Override
     public Permissions getPermissions(final String fileId) {
         JiraAuthenticationContext jiraAuthenticationContext = ComponentAccessor.getJiraAuthenticationContext();
         ApplicationUser user = jiraAuthenticationContext.getLoggedInUser();
@@ -53,7 +74,9 @@ public class ConfigServiceImpl extends DefaultConfigService {
 
         return Permissions.builder()
                 .edit(editPermission && isEditable)
-                .chat(getSettingsManager().getSettingBoolean("customization.chat", true))
+                .protect(user != null)
+                .chat(user != null
+                        && getSettingsManager().getSettingBoolean("customization.chat", true))
                 .build();
     }