+++ src/main/resources/js/editor.js
@@ -14,6 +14,15 @@
         }
     }
 
+    var onMakeActionLink = function (event) {
+        var actionData = JSON.stringify(event.data);
+        var actionLink = new URL(location.href);
+
+        actionLink.searchParams.set("actionData", actionData);
+
+        docEditor.setActionLink(actionLink.toString());
+    };
+
     var onRequestSaveAs = function (event) {
         var url = event.data.url;
         var fileType = event.data.fileType ? event.data.fileType : event.data.title.split(".").pop();
@@ -68,6 +77,43 @@
         });
     };
 
+    const onRequestUsers = function(event) {
+        switch (event.data.c) {
+            case "info":
+                const queryParams = new URLSearchParams({
+                    type: "users-info"
+                });
+
+                fetch(AJS.Meta.get("onlyoffice-api-path") + "?" + queryParams.toString(), {
+                    method: "POST",
+                    headers: {
+                        "Content-Type": "application/json"
+                    },
+                    body: JSON.stringify({
+                        ids: event.data.id
+
+                    })
+                }).then(function(response) {
+                    if (!response.ok) {
+                         return;
+                    }
+
+                    return  response.json();
+                })
+                .then(function (data) {
+                    if (!data) {
+                        return;
+                    }
+
+                    docEditor.setUsers({
+                        "c": event.data.c,
+                        "users": data.users,
+                    });
+                })
+                break;
+        }
+    }
+
     var connectEditor = function() {
         if (typeof DocsAPI === "undefined") {
             alert("ONLYOFFICE is not available. Please contact us at support@onlyoffice.com");
@@ -89,7 +135,9 @@
         }
 
         config.events = {
-            "onAppReady": onAppReady
+            "onAppReady": onAppReady,
+            "onMakeActionLink": onMakeActionLink,
+            "onRequestUsers": onRequestUsers
         };
 
         if (AJS.Meta.get("onlyoffice-can-save-as")) {